refractor: move Volume Renderer components into their own sub-folder
